lxd/entity: Moves `shared/entity` to `lxd/entity`.
The entity package was originally added to shared so that these types
are available to the client. This is an issue if the client is not the
same version as the server. For example, if the server is an older
version than the client, the client may accept an entity type that is
not available on the server yet. Therefore all validation should occur
server-side.
