+++ lxd/entity/type.go
