+++ lxd/entity/type_test.go
